reglage pb saturation + debuguage + cast entier arrondi + changement de type des entiers 32 -> 16

+++ include/decomp.h
@@ -8,15 +8,15 @@
 
 
 extern uint16_t get_indice(struct bitstream *stream, uint32_t nb_bits);
-extern int32_t trad_DC(struct bitstream *stream, struct jpeg_desc *jpeg);
+extern int16_t trad_DC(struct bitstream *stream, struct jpeg_desc *jpeg);
 extern struct symbole_AC trad_AC(struct bitstream *stream, struct jpeg_desc *jpeg);
 extern int32_t valeur_magnitude(uint8_t magnitude, uint16_t);
-extern int32_t *trad_bloc(struct bitstream *stream, struct jpeg_desc *jpeg);
+extern int16_t *trad_bloc(struct bitstream *stream, struct jpeg_desc *jpeg);
 
 struct symbole_AC {
   int EOB;
   int8_t nb_zeros;
-  int32_t valeur;
+  int16_t valeur;
 };
